refactor(*): modify the character search / selection so they can share fragments
