+++ EorzeanInfo/app/src/main/java/chesire/eorzeaninfo/classes/dagger/CharacterStorageComponent.java
@@ -4,7 +4,6 @@ import javax.inject.Singleton;
 
 import chesire.eorzeaninfo.LaunchActivity;
 import chesire.eorzeaninfo.views.CharacterProfileActivity;
-import chesire.eorzeaninfo.views.CharacterSearchDialogFragment;
 import chesire.eorzeaninfo.views.CharacterSelectFragment;
 import dagger.Component;
 
@@ -14,8 +13,6 @@ public interface CharacterStorageComponent {
     void inject(CharacterProfileActivity activity);
 
     void inject(LaunchActivity activity);
-
-    void inject(CharacterSearchDialogFragment fragment);
 
     void inject(CharacterSelectFragment fragment);
 }
