+++ EorzeanInfo/app/src/main/res/layout/activity_character_search.xml
@@ -1,83 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/character_search_container"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <android.support.design.widget.TextInputLayout
-        android:id="@+id/character_search_name_input_layout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="8dp">
-
-        <android.support.design.widget.TextInputEditText
-            android:id="@+id/character_search_name_input"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="@string/search_character_name"
-            android:inputType="textCapWords"
-            android:lines="1"/>
-    </android.support.design.widget.TextInputLayout>
-
-    <android.support.v7.widget.AppCompatTextView
-        android:id="@+id/character_search_data_centre_label"
-        style="@style/SpinnerLabels"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/character_search_name_input_layout"
-        android:layout_marginEnd="8dp"
-        android:layout_marginLeft="8dp"
-        android:layout_marginRight="8dp"
-        android:layout_marginStart="8dp"
-        android:text="@string/search_data_centre_label" />
-
-    <android.support.v7.widget.AppCompatSpinner
-        android:id="@+id/character_search_data_centre_selection"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/character_search_data_centre_label"
-        android:layout_marginBottom="8dp"
-        android:layout_marginLeft="8dp"
-        android:layout_marginRight="8dp" />
-
-    <android.support.v7.widget.AppCompatTextView
-        android:id="@+id/character_search_server_label"
-        style="@style/SpinnerLabels"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/character_search_data_centre_selection"
-        android:layout_marginEnd="8dp"
-        android:layout_marginLeft="8dp"
-        android:layout_marginRight="8dp"
-        android:layout_marginStart="8dp"
-        android:text="@string/search_server_label" />
-
-    <android.support.v7.widget.AppCompatSpinner
-        android:id="@+id/character_search_server_selection"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/character_search_server_label"
-        android:layout_marginBottom="8dp"
-        android:layout_marginLeft="8dp"
-        android:layout_marginRight="8dp" />
-
-    <ProgressBar
-        android:id="@+id/character_search_progress_indicator"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:indeterminate="true"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_margin="8dp"
-        android:visibility="invisible"/>
-
-    <android.support.v7.widget.AppCompatButton
-        android:id="@+id/character_search_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_margin="8dp"
-        android:text="@string/search_search_button" />
-
-</RelativeLayout>
+    android:layout_height="match_parent" />
